restructures code, decouples logic for digestion methods, adds directories for export, import, digestion
