+++ src/digestion/mapping.ts
@@ -1,19 +1,6 @@
 import * as readline from 'readline/promises';
 import { stdin, stdout } from 'node:process';
-import { randomUUID } from 'crypto';
 import { sheets_v4 } from 'googleapis';
-import { readFile, writeFile } from 'fs/promises';
-import * as path from 'path';
-import { ImportEventsBodyParam } from '@api/mixpaneldevdocs/types';
-
-/* The file .mixpanel stores the user's Mixpanel credentials, and is
-created automatically when authorizeMixpanel() is called for the first
-time. If updating user's Mixpanel credentials, simply delete the .mixpanel
-file and run authorizeMixpanel() again.
-*/
-
-// Store the path of the .mixpanel file
-const MIX_CRED = path.join(process.cwd(), '.mixpanel');
 
 export type MixpanelCredentials = {
   PROJECT_ID: string;
@@ -26,109 +13,6 @@ export type MappingType = {
   eventName: string;
   time: string;
   custom: { [key: string]: string };
-};
-
-type MixpanelEventType = {
-  properties: {
-    [x: string]: unknown;
-    time: number;
-    $insert_id: string;
-    distinct_id: string;
-  };
-  event: string;
-};
-
-/* Returns the user's Mixpanel Credentials saved in fs,
-or prompts user for the credentials, saves them to fs, then returns them */
-const authorizeMixpanel = async (): Promise<MixpanelCredentials> => {
-  try {
-    /* Reads previously authorized Mixpanel credentials from the saved
-    .mixpanel file */
-    const savedCredentials = await readFile(MIX_CRED);
-
-    // Store the credentials as an object
-    const credentials = JSON.parse(savedCredentials.toString());
-
-    // Destructure the required properties
-    const { PROJECT_ID, SERVICE_ACCOUNT, SERVICE_ACCOUNT_PASSWORD } = credentials;
-
-    // If all required properties are not null-ish, return the credentials object
-    if (PROJECT_ID && SERVICE_ACCOUNT && SERVICE_ACCOUNT_PASSWORD) {
-      return credentials;
-    }
-    throw new Error();
-  } catch (err) {
-    /* Creates a new .mixpanel file containing user's Mixpanel credentials using
-    CLI user input */
-
-    // Define new object to store user responses from CLI
-    const input: { [key: string]: string } = {};
-
-    // Store a reference to new createInterface instance
-    const rl = readline.createInterface(stdin, stdout);
-
-    // Define function to validate CLI inputs
-    const validate = async (r: string, k: string) => {
-    // Define a variable to store the number of user attempts at a valid response
-      let retries = 3;
-      // Store a reference to the initial response argument
-      let response = r;
-
-      while (!response && retries > 0) {
-      // Repromt user for a valid response
-      // eslint-disable-next-line no-await-in-loop
-        response = await rl.question(
-          'Credentials cannot be empty. Please re-enter\t',
-        );
-
-        // Decrement the variable representing allow attempts by 1
-        retries -= 1;
-      }
-
-      // If a valid response is not entered after 3 attemps, throw error
-      if (retries === 0 && !response) {
-        throw new Error('You have not entered valid responses.\nPlease review your credentials and try running the script again.');
-      }
-
-      input[k] = response;
-    };
-
-    // Define the questions to ask user
-    const questions = {
-      PROJECT_ID: 'What is your Mixpanel Project ID?',
-      SERVICE_ACCOUNT: 'What is your Mixpanel Service Account Username?',
-      SERVICE_ACCOUNT_PASSWORD:
-        'What is your Mixpanel Service Account Password?',
-    } as MixpanelCredentials;
-
-    // Store the keys from the questions (associated with credentials keys)
-    const credKeys = Object.keys(questions);
-
-    // Iterate through the questions, use for loop to await CLI
-    for (let i = 0; i < credKeys.length; i += 1) {
-      // Store the current credential key (see questions object)
-      const credKey = credKeys[i];
-
-      // Await user response
-      // eslint-disable-next-line no-await-in-loop
-      const unsafeResponse = await rl.question(
-        `${questions[credKey as keyof typeof questions]} \t`,
-      );
-
-      // Await validation of the response, if valid input object will be updated
-      // eslint-disable-next-line no-await-in-loop
-      await validate(unsafeResponse, credKey);
-    }
-
-    // Close the readline interface
-    rl.close();
-
-    // Await writing the credentials to fs
-    await writeFile(MIX_CRED, JSON.stringify(input));
-
-    // Return the CLI input as the credentials
-    return input as typeof questions;
-  }
 };
 
 /* Gets user's Google Spreadsheet credentials for a specific spreadsheet */
@@ -322,52 +206,7 @@ const createMap = async (columns: Set<string>): Promise<MappingType> => {
   return map;
 };
 
-const digest = (data: string[][], map: MappingType) => {
-  // Define an empty obejct to serve as a map of indexes
-  const hash: { [key: string]: number } = {};
-
-  // Store the header row for reference
-  const headerRow = data[0];
-
-  // Iterate over each col in header row, and add its index to the hash map
-  headerRow.forEach((col, index) => {
-    hash[col] = index;
-  });
-
-  // Define an array of events that will be returned
-  const events: ImportEventsBodyParam = [];
-
-  // Iterate over all not header rows
-  for (let i = 1; i < data.length; i += 1) {
-    // Copy the row reference
-    const row = data[i];
-
-    // Define the event object with the required fields
-    const eventObject: MixpanelEventType = {
-      event: row[hash[map.eventName]],
-      properties: {
-        $insert_id: randomUUID(),
-        distinct_id: row[hash[map.distinct_id]],
-        time: (map.time !== '' ? row[hash[map.time]] as unknown as number : Date.now()),
-      },
-    };
-
-    // Add the custom properties to the event object
-    Object.entries(map.custom).forEach(([prop, col]) => {
-      eventObject.properties[prop] = row[hash[col]];
-    });
-
-    // Push the current event object into the events array
-    events.push(eventObject);
-  }
-
-  // Return the events array
-  return events;
-};
-
 export {
-  authorizeMixpanel,
   createMap,
-  digest,
   getSpreadsheet,
 };
