+++ tests/mapping.test.ts
@@ -1,16 +1,9 @@
-/* eslint-disable @typescript-eslint/naming-convention */
 import { createInterface } from 'readline/promises';
-import { join } from 'path';
-import * as fsp from 'fs/promises';
-import * as fspMockModule from '../__mocks__/fs/promises';
-import { authorizeMixpanel, createMap, getSpreadsheet } from '../src/digest/mapping';
+import { createMap, getSpreadsheet } from '../src/digestion/mapping';
 
 // Mock dependencies
 jest.mock('readline/promises');
 jest.mock('fs/promises');
-
-// Declare mock types
-type FSPMockType = typeof fspMockModule & typeof fsp;
 
 // Declare label for mocked readline interface constructor
 const mockedCreateInterface = createInterface as jest.Mock;
@@ -44,145 +37,6 @@ describe('getSpreadsheet', () => {
   });
 });
 
-describe('authorizeMixpanel', () => {
-  // Store internal mock module method for updating mocked file system
-  const { __setMockFiles } = fsp as FSPMockType;
-
-  // Save the current work directory
-  const filepath = join(process.cwd(), '.mixpanel');
-
-  // Save empty credentials to be used in tests
-  const emptyCredentials = `{
-    "PROJECT_ID": "",
-    "SERVICE_ACCOUNT": "",
-    "SERVICE_ACCOUNT_PASSWORD": ""
-  }`;
-
-  beforeEach(async () => {
-    // Clear mocked methods before each test
-    mockedCreateInterface.mockReset();
-
-    // Overwrite the mocked file system storage
-    __setMockFiles({ './emptyPath': '' });
-  });
-
-  test('returns saved credentials object', async () => {
-    // Save empty object to be used as mocked file system storage
-    const mockFiles: { [key:string]: string } = {};
-
-    // Save test credentials to filepath in the storage object
-    mockFiles[filepath] = `{
-      "PROJECT_ID": "savedProjectId",
-      "SERVICE_ACCOUNT": "savedServiceAccount",
-      "SERVICE_ACCOUNT_PASSWORD": "savedServiceAccountPassword"
-    }`;
-
-    // Save mocked credential file in the mocked file system
-    __setMockFiles(mockFiles);
-
-    // Get credentials object
-    const credentials = await authorizeMixpanel();
-
-    // Reads credential object saved in file system
-    expect(fsp.readFile).toBeCalledWith(filepath);
-
-    // Object should be returned when promise resolves
-    expect(credentials).toBeInstanceOf(Object);
-
-    // Object should have the appropriate shape (Mixpanel Credentials)
-    expect(credentials).toEqual(expect.objectContaining({
-      PROJECT_ID: 'savedProjectId',
-      SERVICE_ACCOUNT: 'savedServiceAccount',
-      SERVICE_ACCOUNT_PASSWORD: 'savedServiceAccountPassword',
-    }));
-  });
-
-  test('triggers CLI input if credentials file does not exist', async () => {
-    // Implement mocked readline for user input
-    mockedCreateInterface.mockReturnValue({
-      question: jest.fn().mockImplementation((): Promise<string> => Promise.resolve('test')),
-      close: jest.fn(),
-    });
-
-    // Call method to ensure logic completes
-    await authorizeMixpanel();
-
-    // Triggers CLI prompts for credentials if file is not saved
-    expect(mockedCreateInterface().question).toBeCalledTimes(3);
-  });
-
-  test('triggers CLI input if credentials file has invalid values', async () => {
-    // Implement mocked readline for user input
-    mockedCreateInterface.mockReturnValue({
-      question: jest.fn().mockImplementation((): Promise<string> => Promise.resolve('test')),
-      close: jest.fn(),
-    });
-
-    // Save empty credentials in the mocked file system
-    __setMockFiles({ './mixpanel': emptyCredentials });
-
-    // Call method to ensure logic completes
-    await authorizeMixpanel();
-
-    // Triggers CLI prompts for credentials if file is not saved
-    expect(mockedCreateInterface().question).toBeCalledTimes(3);
-  });
-
-  test('prompts for CLI input again if input has invalid values', async () => {
-    // Implement mocked readline for user input
-    mockedCreateInterface.mockReturnValue({
-      question: jest.fn()
-        .mockImplementationOnce((): Promise<string> => Promise.resolve(''))
-        .mockImplementationOnce((): Promise<string> => Promise.resolve('mockProjectId'))
-        .mockImplementationOnce((): Promise<string> => Promise.resolve(''))
-        .mockImplementationOnce((): Promise<string> => Promise.resolve('mockServiceAccount'))
-        .mockImplementationOnce((): Promise<string> => Promise.resolve(''))
-        .mockImplementationOnce((): Promise<string> => Promise.resolve('mockServiceAccountPassword')),
-      close: jest.fn(),
-    });
-
-    // Save empty credentials in the mocked file system
-    __setMockFiles({ './mixpanel': emptyCredentials });
-
-    // Call method to ensure logic completes
-    await authorizeMixpanel();
-
-    // Triggers CLI prompts for credentials again if empty string received
-    expect(mockedCreateInterface().question).toBeCalledTimes(6);
-  });
-
-  test('generates credentials object from CLI input', async () => {
-    // Implement mocked readline for user input
-    mockedCreateInterface.mockReturnValue({
-      question: jest.fn()
-        .mockImplementationOnce((): Promise<string> => Promise.resolve('mockProjectId'))
-        .mockImplementationOnce((): Promise<string> => Promise.resolve('mockServiceAccount'))
-        .mockImplementationOnce((): Promise<string> => Promise.resolve('mockServiceAccountPassword')),
-      close: jest.fn().mockImplementation(() => undefined),
-    });
-
-    // Get credentials object
-    const credentials = await authorizeMixpanel();
-
-    // Saves credential object to file system
-    expect(fsp.writeFile).toBeCalledWith(filepath, JSON.stringify({
-      PROJECT_ID: 'mockProjectId',
-      SERVICE_ACCOUNT: 'mockServiceAccount',
-      SERVICE_ACCOUNT_PASSWORD: 'mockServiceAccountPassword',
-    }));
-
-    // Object should be returned when promise resolves
-    expect(credentials).toBeInstanceOf(Object);
-
-    // Object should have the appropriate shape (Mixpanel Credentials)
-    expect(credentials).toEqual(expect.objectContaining({
-      PROJECT_ID: 'mockProjectId',
-      SERVICE_ACCOUNT: 'mockServiceAccount',
-      SERVICE_ACCOUNT_PASSWORD: 'mockServiceAccountPassword',
-    }));
-  });
-});
-
 describe('create mapping from CLI input', () => {
   const headerRow = [
     'eventName', 'hasDeclaredMajor', 'major', 'studentName', 'eventId',
@@ -206,7 +60,6 @@ describe('create mapping from CLI input', () => {
 
     // Await mapping object, created with above mock input
     const map = await createMap(new Set(headerRow));
-    console.log(map);
 
     // First user input should be ignored, and user reprompted with question
     expect(Object.keys(map)).not.toContain('fakeName');
