+++ tests/mocks/mockReadline.test.ts
